pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tacbot"
version = "1.0.0"
description = "Discord bot for Havoc Tactical Group"
authors = [{name = "Liam", email = "liamcjackson@pm.me"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "discord.py>=2.6.0",
    "google-auth>=2.40.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.182.0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio", 
    "pytest-mock",
    "pytest-cov",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock", 
    "pytest-cov",
    "black",
    "flake8",
    "pre-commit"
]

[project.urls]
"Homepage" = "https://github.com/ljackson330/tacbot"
"Bug Reports" = "https://github.com/ljackson330/tacbot/issues"
"Source" = "https://github.com/ljackson330/tacbot"

[tool.setuptools.packages.find]
where = ["."]
include = ["cogs*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=cogs --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"

[tool.black]
line-length = 150
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''